requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-baton==2.2.3
django-ckeditor==6.1.0
django-debug-toolbar==3.1.1
django-extensions==3.1.5
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
numpy==1.21.4
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.1
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
